+++ src/main/java/cyou/lositha/springehcacheexample/employeeCrud/dao/DesignationRepository.java
@@ -1,9 +1,10 @@
 package cyou.lositha.springehcacheexample.employeeCrud.dao;
 
-import cyou.lositha.springehcacheexample.employeeCrud.entity.Designation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import cyou.lositha.springehcacheexample.employeeCrud.entity.Designation;
+
 @Repository
-public interface DesignationRepository extends JpaRepository<Designation, Integer> {
+public interface DesignationRepository extends JpaRepository<Designation, Long> {
 }
