+++ src/main/java/cyou/lositha/springehcacheexample/employeeCrud/dao/EmployeeRepository.java
@@ -1,9 +1,10 @@
 package cyou.lositha.springehcacheexample.employeeCrud.dao;
 
-import cyou.lositha.springehcacheexample.employeeCrud.entity.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import cyou.lositha.springehcacheexample.employeeCrud.entity.Employee;
+
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> {
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
 }
