+++ src/main/java/cyou/lositha/springehcacheexample/employeeCrud/dao/GenderRepository.java
@@ -1,9 +1,10 @@
 package cyou.lositha.springehcacheexample.employeeCrud.dao;
 
-import cyou.lositha.springehcacheexample.employeeCrud.entity.Gender;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import cyou.lositha.springehcacheexample.employeeCrud.entity.Gender;
+
 @Repository
-public interface GenderRepository extends JpaRepository<Gender, Integer> {
+public interface GenderRepository extends JpaRepository<Gender, Long> {
 }
