+++ src/main/java/cyou/lositha/springehcacheexample/employeeCrud/dao/HobbiesRepository.java
@@ -1,9 +1,10 @@
 package cyou.lositha.springehcacheexample.employeeCrud.dao;
 
-import cyou.lositha.springehcacheexample.employeeCrud.entity.Hobbies;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import cyou.lositha.springehcacheexample.employeeCrud.entity.Hobbies;
+
 @Repository
-public interface HobbiesRepository extends JpaRepository<Hobbies, Integer> {
+public interface HobbiesRepository extends JpaRepository<Hobbies, Long> {
 }
