+++ src/main/java/cyou/lositha/springehcacheexample/employeeCrud/dao/WeekDayRepository.java
@@ -1,9 +1,10 @@
 package cyou.lositha.springehcacheexample.employeeCrud.dao;
 
-import cyou.lositha.springehcacheexample.employeeCrud.entity.WeekDay;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import cyou.lositha.springehcacheexample.employeeCrud.entity.WeekDay;
+
 @Repository
-public interface WeekDayRepository extends JpaRepository<WeekDay, Integer> {
+public interface WeekDayRepository extends JpaRepository<WeekDay, Long> {
 }
